feat: Add conditional check for graph representation in Blip2OPT
This commit adds a conditional check for the "graph" representation style in the `Blip2OPT` class in the `blip2_opt.py` file. The check ensures that the `prompt_embeds` are updated with the `mol_tokens` only if the "graph" representation is specified in the `mol_representation` argument. This improves the flexibility of molecule representation in the model.