feat: Add gnn_jk and used_gnn_layer arguments to Blip2Stage1
This commit adds the `gnn_jk` and `used_gnn_layer` arguments to the `Blip2Stage1` class in the `blip2_stage1.py` file. The `gnn_jk` argument allows the user to specify the GNN JK operation type, with options for 'concat', 'max', or 'mean'. The `used_gnn_layer` argument allows the user to specify the index of the GNN layer to be used. These additions provide more control over the GNN operations and layer selection in the model.