refactor: Update Stage3DM loader logic for multi-task mode
This refactor updates the loader logic in the `Stage3DM` class of `stage3_dm.py` to handle the multi-task mode. The previous implementation only supported the `ft` mode, but now it has been extended to handle the `multi_task` mode as well. The loader now checks the value of the `root` attribute and creates the appropriate data loader based on the mode. If `root` is set to "multi_task", multiple data loaders are created for each task in the `concat_datasets` dictionary. Otherwise, a single data loader is created for the `train_dataset`. This change improves the flexibility and functionality of the `Stage3DM` class in handling different modes of operation.