refactor & feat: Update data_provider classes to use sliced lists. add get_necessary_data to get reaction data instance
This commit refactors the `MolInstructionDatset`, `ChEBIDatset`, and `SMolInstructDataset` classes in the `data_provider` module. The changes involve updating the `set_necesary_data` method to use sliced lists for the `input_list`, `label_list`, and `instruction_list` attributes. This ensures that the original data lists are not modified during the data processing.

Refactoring the code in this way improves code readability and maintainability.